Use int64 instead of int32 to store the histogram bars.
Int32 might overflow so just to be safe we use int64.

+++ dae/dae/genomic_resources/score_statistics.py
@@ -44,7 +44,7 @@ class Histogram:
 
         self.y_scale = y_scale
 
-        self.bars = np.zeros(bins_count, dtype=np.int32)
+        self.bars = np.zeros(bins_count, dtype=np.int64)
 
     @staticmethod
     def from_config(conf):
@@ -76,7 +76,7 @@ class Histogram:
         )
 
         hist.bins = np.array(d["bins"])
-        hist.bars = np.array(d["bars"], dtype=np.int32)
+        hist.bars = np.array(d["bars"], dtype=np.int64)
 
         return hist
